+++ src/MainApps.hs
@@ -42,8 +42,9 @@ ampersandWorker eGlobalRun = do
     Right (globalMonoid,run) -> do
       global <- globalOptsFromMonoid isTerminal defaultOuptutDir globalMonoid
       -- when (globalLogLevel global == LevelDebug) $ hPutStrLn stderr versionString'
+      
       withRunnerGlobal global $ run `catch` \e ->
-          -- This special handler stops "stack: " from being printed before the
+          -- This special handler stops "ampersand: " from being printed before the
           -- exception
           case fromException e of
               Just ec -> exitWith ec